refactor: branch name and replace domain
